- initial support for navigation compose - fix for material theming on child view controllers - updated sample to demonstrate both voyager and navigation compose
