+++ composeApp/src/commonMain/kotlin/com/kashif/sample/voyager/SampleBottomSheet.kt
@@ -1,0 +1,36 @@
+package com.kashif.sample.voyager
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import cafe.adriel.voyager.core.screen.Screen
+import cafe.adriel.voyager.navigator.bottomSheet.LocalBottomSheetNavigator
+import com.kashif.voyant.hideX
+
+class SampleBottomSheet : Screen {
+    @Composable
+    override fun Content() {
+
+        val bottomSheetNavigator = LocalBottomSheetNavigator.current
+        Column(
+            modifier = Modifier.fillMaxWidth(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = "Bottom Sheet",
+                style = MaterialTheme.typography.headlineSmall
+            )
+            Button(
+                onClick = { bottomSheetNavigator.hideX() },
+                content = { Text("Close") }
+            )
+        }
+    }
+}
