+++ composeApp/src/commonMain/kotlin/com/kashif/sample/voyager/ScreenB.kt
@@ -1,0 +1,36 @@
+package com.kashif.sample.voyager
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import cafe.adriel.voyager.core.screen.Screen
+import cafe.adriel.voyager.navigator.LocalNavigator
+import cafe.adriel.voyager.navigator.currentOrThrow
+import com.kashif.voyant.pushX
+
+class ScreenB : Screen {
+    @Composable
+    override fun Content() {
+        val navigator = LocalNavigator.currentOrThrow
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text ="Screen B",
+                style = MaterialTheme.typography.headlineSmall
+            )
+            Button(
+                onClick = { navigator.pushX(screen = ScreenC()) },
+                content = { Text("Go to Screen C") }
+            )
+        }
+    }
+}
