+++ navigationComposeX/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
     alias(libs.plugins.compose.compiler)
     alias(libs.plugins.compose)
     id("com.vanniktech.maven.publish") version "0.28.0"
+    kotlin("plugin.serialization") version "2.0.20"
 }
 
 group = "com.kashif.voyant.navigationcompose"
