+++ navigationComposeX/src/appleMain/kotlin/com/kashif/voyant_navigation_compose/Extensions.apple.kt
@@ -1,38 +1,17 @@
 package com.kashif.voyant_navigation_compose
 
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ExperimentalComposeApi
+import androidx.compose.runtime.remember
 import androidx.compose.ui.uikit.OnFocusBehavior
 import androidx.compose.ui.window.ComposeUIViewController
-import androidx.navigation.NavHostController
+import com.kashif.voyant_common.ThemeManager
 import com.kashif.voyant_common.extensions.UIViewControllerWrapper
 import com.kashif.voyant_common.extensions.getNavigationController
 import platform.Foundation.NSLog
+import platform.UIKit.UINavigationController
 import platform.UIKit.UIViewController
 import platform.UIKit.hidesBottomBarWhenPushed
-
-actual fun NavHostController.popBackStackX() {
-    val navigationController = getNavigationController()
-    navigationController?.let { navController ->
-        if (navController.viewControllers.size > 1) {
-            navController.popViewControllerAnimated(true)
-        } else {
-            NSLog("Cannot pop. Only one view controller in the stack.")
-        }
-    } ?: run {
-        NSLog("NavigationController is null")
-    }
-}
-
-
-actual fun NavHostController.navigateX(route: VoyantRoute) {
-
-    val viewController = extendedComposeViewController(screen = route)
-    viewController.hidesBottomBarWhenPushed = true
-    val navigationController = getNavigationController()
-    navigationController?.pushViewController(viewController, animated = true) ?: run {
-        NSLog("NavigationController is null")
-    }
-}
 
 
 @OptIn(ExperimentalComposeApi::class)
@@ -45,9 +24,43 @@ fun extendedComposeViewController(
         onFocusBehavior = OnFocusBehavior.DoNothing
         opaque = isOpaque
     }) {
-
-        screen.content()
+        ThemeManager.currentTheme {
+            screen.content()
+        }
     }
-
     return UIViewControllerWrapper(uiViewController)
 }
+
+class IOSNavController(
+    private val navController: UINavigationController?,
+) : NavigationController {
+    override fun navigate(route: VoyantRoute) {
+        val viewController =
+            extendedComposeViewController(screen = route)
+        viewController.hidesBottomBarWhenPushed = true
+        navController?.pushViewController(viewController, animated = true) ?: run {
+            NSLog("NavigationController is null")
+        }
+    }
+
+    override fun popBackStack() {
+        navController?.let { navController ->
+            if (navController.viewControllers.size > 1) {
+                navController.popViewControllerAnimated(true)
+            } else {
+                NSLog("Cannot pop. Only one view controller in the stack.")
+            }
+        } ?: run {
+            NSLog("NavigationController is null")
+        }
+    }
+}
+
+@Composable
+actual fun rememberKMPNavController(): NavigationController {
+    val navController = getNavigationController()
+    return remember(navController) {
+        IOSNavController(navController)
+    }
+}
+
