+++ navigationComposeX/src/commonMain/kotlin/com/kashif/voyant_navigation_compose/Extensions.kt
@@ -1,15 +1,20 @@
 package com.kashif.voyant_navigation_compose
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
 
-
-
-expect fun NavHostController.popBackStackX()
-
-expect fun NavHostController.navigateX(route: VoyantRoute)
+@Composable
+expect fun rememberKMPNavController(): NavigationController
 
 interface VoyantRoute{
     @Composable
     fun content()
 }
+
+
+interface NavigationController {
+    fun navigate(route: VoyantRoute)
+    fun popBackStack()
+
+}
+
+
